requirements.txt
# Core Framework
Flask>=2.3.0,<3.0.0
Werkzeug>=2.3.0,<3.0.0

# AI and Machine Learning
openai>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0

# Database
# SQLite3 is included with Python

# HTTP and API
urllib3>=2.0.0,<3.0.0
certifi>=2023.0.0

# Development and Testing
pytest>=7.0.0,<8.0.0
pytest-cov>=4.0.0,<5.0.0
black>=23.0.0,<24.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.0.0,<2.0.0

# Code Quality
pre-commit>=3.0.0,<4.0.0

# Production
gunicorn>=21.0.0,<22.0.0

# Environment Management
python-dotenv>=1.0.0,<2.0.0

# Logging
colorlog>=6.0.0,<7.0.0

# Security
cryptography>=41.0.0,<42.0.0 